requirements.txt
edge-tts>=6.1.0
pyinstaller>=6.0.0
Pillow>=10.0.0
